Move UserProfileSettings to layout folder
Also remove a useless inclusion of that component in UserMenuLoginForm
